% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_IV_7.R
\name{clean_IV_7}
\alias{clean_IV_7}
\title{Package for IV-7 Form from Statistics Bureau database, showing weekly hours of work by husband and wife  - fix this}
\usage{
clean_IV_7(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Package for IV-7 Form from Statistics Bureau database, showing weekly hours of work by husband and wife  - fix this
}
