% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_IV_9.R
\name{clean_IV_9}
\alias{clean_IV_9}
\title{Reads in and organizes Form IV-9 from Japan's Statistics Bureau}
\usage{
clean_IV_9(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Reads in and organizes Form IV-9 from Japan's Statistics Bureau
}
